how did i code this with a straight face?

+++ Cryptographer/Decoders/Generic/Base45.cs
@@ -8,10 +8,10 @@ namespace Cryptographer.Decoders
     {
         private const string Base45Characters = MethodDictionaries.Base45Characters;
 
-        private int CharToVal(char c)
+        private byte CharToVal(char c)
         {
-            sbyte i = (sbyte)Base45Characters.IndexOf(c);
-            return (i == -1 ? 255 : i);
+            int i = Base45Characters.IndexOf(c);
+            return (byte)(i == -1 ? 255 : i);
         }
         private bool TryFromBase45String(string input, out byte[]? output)
         {
@@ -25,9 +25,9 @@ namespace Cryptographer.Decoders
             for (int i = 0; i < input.Length; i += 3)
             {
                 //n=c+(d×45)+(e×45^2)
-                byte c = (byte)CharToVal(input[i]);
-                byte d = (byte)CharToVal(input[i + 1]);
-                byte e = (byte)CharToVal(input[i + 2]);
+                byte c = CharToVal(input[i]);
+                byte d = CharToVal(input[i + 1]);
+                byte e = CharToVal(input[i + 2]);
                 if (c == 255 || d == 255 || e == 255) return false; // aka failed in chartoval
 
                 int n = c + (d * 45) + (e * 45 * 45);
